apps/backend/requirements.txt

fastapi
uvicorn[standard]
numpy
scipy
soundfile
python-multipart
aiofiles
faiss-cpu
